use transactional annotation and fix delete

+++ src/main/java/testing/model/dao/UserDAO.java
@@ -2,6 +2,7 @@ package testing.model.dao;
 
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
+import javax.transaction.Transactional;
 
 import testing.model.entity.User;
 
@@ -13,18 +14,21 @@ public class UserDAO implements IUserDAO {
 	public User getById(int id) {
 		return entityManager.find(User.class, id);
 	}
-
+	
+	@Transactional
 	public void add(User user) {
 		entityManager.persist(user);
 	}
 
+	@Transactional
 	public void update(User user) {
 		entityManager.merge(user);
 	}
 
-
+	@Transactional
 	public void delete(int id) {
-		User user = entityManager.find(User.class, id);
+		User user = getById(id);
+		if(user != null)
 		entityManager.remove(user);
 	}
 }
